+++ lib/fp_util.dart
@@ -1,3 +1,4 @@
+export 'package:gap/gap.dart';
 export 'package:image_picker/image_picker.dart' show XFile;
 
 export 'src/constants/constants.dart';